+++ package.json
@@ -52,7 +52,7 @@
     "ipfs-repo": "~0.17.0",
     "ipld-dag-cbor": "~0.11.1",
     "ipld-dag-pb": "~0.11.2",
-    "ipld-eth-star": "^1.4.1",
+    "ipld-ethereum": "^1.4.2",
     "ipld-git": "^0.1.0",
     "is-ipfs": "~0.3.0",
     "lodash.flatten": "^4.4.0",
