+++ test/browser.js
@@ -42,6 +42,6 @@ describe('Browser', () => {
   require('./ipld-dag-cbor')(repo)
   require('./ipld-git')(repo)
   require('./ipld-eth-block')(repo)
-  require('./ipld-eth-star')(repo)
+  require('./ipld-eth')(repo)
   require('./ipld-all')
 })
